pyproject.toml
[tool.poetry]
name = "mine-workspace"
version = "0.1.0"
description = ""
authors = ["纯二 <chuxueonline@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
fastapi = {extras = ["all"], version = "^0.115.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
